Functional Flowchart Controller
Changed several variables to properly create a Flowchart.

+++ src/main/java/base/studentFlowchartDb/Flowchart.java
@@ -16,13 +16,13 @@ public class Flowchart implements Serializable {
 
     @Id
     private String id;
-    private Long studentId;
+    private int studentId;
     private String name;
     private List<PlannedCourse> plannedCourses = new ArrayList<PlannedCourse>();
 
     public Flowchart() {}
 
-    public Flowchart(Long studentId, String name, List<PlannedCourse> plannedCourses) {
+    public Flowchart(int studentId, String name, List<PlannedCourse> plannedCourses) {
         this.studentId = studentId;
         this.name = name;
         this.plannedCourses.addAll(plannedCourses);
@@ -36,11 +36,11 @@ public class Flowchart implements Serializable {
         this.id = id;
     }
 
-    public Long getStudentId() {
+    public int getStudentId() {
         return studentId;
     }
 
-    public void setStudentId(Long studentId) {
+    public void setStudentId(int studentId) {
         this.studentId = studentId;
     }
